+++ engine/execution/ingestion/block_queue/queue.go
@@ -1,4 +1,4 @@
-package blocks
+package block_queue
 
 import (
 	"fmt"
